Convert "Loading" page to typescript
