+++ src/pages/loading/styles.ts
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+export const LoadScreenWrap = styled.div`
+  height: 100vh;
+  width: 100vw;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+
+  img {
+    width: 150px;
+  }
+
+  span {
+    margin-top: 16px;
+  }
+`;
